Move EncryptedStringField-specific tests to tests.mongo.fields.test_crypto
