+++ tests/test_legacy.py
@@ -1,21 +1,16 @@
 # -*- coding: utf-8 -*-
 
-import random
-import string
 import unittest
 
 from flask import Flask
-from mongoengine import connection
 from mongoengine.errors import DoesNotExist
 from mongoengine.fields import (ReferenceField, SafeReferenceField,
                                 SafeReferenceListField, StringField,
                                 IntField)
 
 from flask_mongoengine import MongoEngine
-from flask_common.crypto import aes_generate_key
 from flask_common.declenum import DeclEnum
 from flask_common.mongo import (
-    EncryptedStringField,
     LowerEmailField,
     LowerStringField,
     custom_query_counter,
@@ -38,71 +33,12 @@ app.config.update(
 db = MongoEngine(app)
 
 
-class Secret(db.Document):
-    password = EncryptedStringField(aes_generate_key())
-
-
 class Book(db.Document):
     pass
 
 
 class Author(db.Document):
     books = SafeReferenceListField(ReferenceField(Book))
-
-
-class SecretTestCase(unittest.TestCase):
-    def test_encrypted_field(self):
-        col = connection._get_db().secret
-
-        # Test creating password
-        s = Secret.objects.create(password='hello')
-        self.assertEqual(s.password, 'hello')
-        s.reload()
-        self.assertEqual(s.password, 'hello')
-
-        cipher = col.find({'_id': s.id})[0]['password']
-        self.assertTrue('hello' not in cipher)
-        self.assertTrue(len(cipher) > 16)
-
-        # Test changing password
-        s.password = 'other'
-        s.save()
-        s.reload()
-        self.assertEqual(s.password, 'other')
-
-        other_cipher = col.find({'_id': s.id})[0]['password']
-        self.assertTrue('other' not in other_cipher)
-        self.assertTrue(len(other_cipher) > 16)
-        self.assertNotEqual(other_cipher, cipher)
-
-        # Make sure password is encrypted differently if we resave.
-        s.password = 'hello'
-        s.save()
-        s.reload()
-        self.assertEqual(s.password, 'hello')
-
-        new_cipher = col.find({'_id': s.id})[0]['password']
-        self.assertTrue('hello' not in new_cipher)
-        self.assertTrue(len(new_cipher) > 16)
-        self.assertNotEqual(new_cipher, cipher)
-        self.assertNotEqual(other_cipher, cipher)
-
-        # Test empty password
-        s.password = None
-        s.save()
-        s.reload()
-        self.assertEqual(s.password, None)
-
-        raw = col.find({'_id': s.id})[0]
-        self.assertTrue('password' not in raw)
-
-        # Test passwords of various lengths
-        for pw_len in range(1, 50):
-            pw = ''.join(random.choice(string.ascii_letters + string.digits) for x in range(pw_len))
-            s = Secret(password=pw)
-            s.save()
-            s.reload()
-            self.assertEqual(s.password, pw)
 
 
 class SafeReferenceListFieldTestCase(unittest.TestCase):
